+++ onnxscript/function_libs/torch_aten/graph_building_test.py
@@ -11,8 +11,10 @@ import onnxscript.testing
 from onnxscript import FLOAT, evaluator
 from onnxscript import opset17 as op
 from onnxscript.function_libs.torch_aten import graph_building, ops
+from onnxscript.tests.common import version_utils
 
 
+@unittest.skipIf(version_utils.torch_older_than("2.0"), "torchscript in 1.13 not supported")
 class TestTorchScriptTracingEvaluator(unittest.TestCase):
     def setUp(self):
         self.opset_version = 17
